mv x/auth.users query
