+++ apps/forms/app/(api)/private/editor/x-supabase/projects/[supabase_project_id]/x/auth.users/query/route.ts
@@ -10,10 +10,16 @@ interface Context {
   };
 }
 
+// [EXTRA SECURITY REQUIRED]
+// Special route for fetching auth.users via proxy supabase client, for in-editor use.
+// this route is protected via supabase_project access RLS.
 export async function GET(req: NextRequest, context: Context) {
   const cookieStore = cookies();
   const supabase = createRouteHandlerXSBClient(cookieStore);
   const { supabase_project_id } = context.params;
+
+  const _q_limit = req.nextUrl.searchParams.get("limit");
+  const limit = _q_limit ? parseInt(_q_limit) : undefined;
 
   // [REQUIRED] RLS gate
   const { data: supabase_project } = await supabase
@@ -26,9 +32,17 @@ export async function GET(req: NextRequest, context: Context) {
     return notFound();
   }
 
-  const client = await createXSupabaseClient(supabase_project_id, {
+  const xclient = await createXSupabaseClient(supabase_project_id, {
     service_role: true,
+    db: {
+      schema: "auth",
+    },
   });
 
-  return NextResponse.json(await client.storage.listBuckets());
+  const res = await xclient.auth.admin.listUsers({
+    page: 1,
+    perPage: limit,
+  });
+
+  return NextResponse.json(res);
 }
