+++ apps/forms/app/(workbench)/[org]/[proj]/[id]/data/(data)/x/auth.users/page.tsx
@@ -25,7 +25,7 @@ export default function XTablePage() {
   const [state, dispatch] = useEditorState();
 
   const {
-    form,
+    supabase_project,
     datagrid_rows_per_page,
     datagrid_orderby,
     datagrid_table_refresh_key,
@@ -40,9 +40,9 @@ export default function XTablePage() {
     });
   }, [datagrid_rows_per_page, datagrid_orderby, datagrid_table_refresh_key]);
 
-  const request = state.connections.supabase?.main_supabase_table_id
-    ? PrivateEditorApi.XSupabase.url_table_auth_users_query(
-        form.form_id,
+  const request = supabase_project
+    ? PrivateEditorApi.XSupabase.url_x_auth_users_get(
+        supabase_project.id,
         serachParams
       )
     : null;
